* commit '5a2526ac629802ff22f71d516f7979f1f4a4c34d':
  Fix useconds_t for LP64.

+++ libc/include/sys/types.h
@@ -98,7 +98,7 @@ typedef __kernel_ssize_t ssize_t;
 typedef __kernel_time_t time_t;
 
 typedef __kernel_suseconds_t suseconds_t;
-typedef unsigned long useconds_t;
+typedef unsigned int useconds_t;
 
 typedef __kernel_daddr_t	daddr_t;
 typedef __kernel_timer_t	timer_t;
